requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
Pillow==5.2.0
requests==2.19.1
six==1.11.0
urllib3==1.23
